+++ src/backend/core/api/viewsets.py
@@ -304,7 +304,7 @@ class RoomViewSet(
         methods=["post"],
         url_path="start-recording",
         permission_classes=[
-            permissions.HasPrivilegesOnRoom,
+            permissions.HasRecordingPermission,
         ],
     )
     @FeatureFlag.require("recording")
@@ -354,7 +354,7 @@ class RoomViewSet(
         methods=["post"],
         url_path="stop-recording",
         permission_classes=[
-            permissions.HasPrivilegesOnRoom,
+            permissions.HasRecordingPermission,
         ],
     )
     @FeatureFlag.require("recording")